Enregistrer les commandes en base de données

+++ app/Http/Controllers/CheckoutController.php
@@ -2,11 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use DateTime;
+use App\Order;
 use Stripe\Stripe;
 use Stripe\PaymentIntent;
-use Illuminate\Http\Request;
 use Illuminate\Support\arr;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
 use Gloudemans\Shoppingcart\Facades\Cart;
+
 
 class CheckoutController extends Controller
 {
@@ -61,10 +65,56 @@ class CheckoutController extends Controller
     public function store(Request $request)
     {
         
-        Cart::destroy();
+        
         $data = $request->json()->all();
-        return $data['paymentIntent'];
+        
+        $order = new Order();
+
+        $order->payment_intent_id = $data['paymentIntent']['id'];
+        $order->amount = $data['paymentIntent']['amount'];
+        $order->payment_create_at = (new DateTime())
+                                    ->setTimestamp($data['paymentIntent']['created'])
+                                    ->format('Y-m-d H:i:s');
+
+        // recuperer les produits 
+
+        $products = [];
+        $i = 0;
+
+        foreach(Cart::content() as $product){
+        $products['product_'.$i][] = $product->model->title;
+        $products['product_'.$i][] = $product->model->price;
+        $products['product_'.$i][] = $product->qty;
+        $i++;
+
+        }
+        $order->products = serialize($products);
+
+        $order->user_id = 15;
+
+        $order->save();
+
+        // verifiaction du paymentIntent 
+
+        if($data['paymentIntent']['status'] == 'succeeded'){
+            Cart::destroy();
+            Session::flash('success','Votre commande à été traiter avec success');
+            return response()->json(['success','Payment Intent Succeeded']);
+        }else{
+            return response()->json(['error','Payment Intent not  Succeeded']);
+        }
+
+        //return $data['paymentIntent'];
        
+    }
+
+    public function thank(){
+
+        if (Session::has('success')) {
+           return view('checkout.thank');
+        } else {
+            return back();
+        }
     }
 
     /**
